requirements.txt
# SPDX-FileCopyrightText: 2022 Alec Delaney, for Adafruit Industries
#
# SPDX-License-Identifier: Unlicense

Adafruit-Blinka
adafruit-circuitpython-busdevice
typing-extensions~=4.0
adafruit-circuitpython-typing~=1.3
